Created roles,permissions and new exception

+++ src/main/java/com/example/Test/controller/UserController.java
@@ -1,13 +1,16 @@
 package com.example.Test.controller;
 
-import com.example.Test.entity.UserEntity;
-import com.example.Test.exeption.UserAlreadyExistException;
+import com.example.Test.exeption.PermissionDeniedException;
 import com.example.Test.exeption.UserNotFoundException;
+import com.example.Test.security.AppUserDetails;
 import com.example.Test.service.UserService;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+
 
 @RestController
 @RequestMapping("/users")
@@ -15,6 +18,8 @@ public class UserController {
 
     @Autowired
     private UserService userService;
+
+    private AppUserDetails userPrincipal;
 
     @GetMapping
     public ResponseEntity getOneUser(@RequestParam String username) {
@@ -29,13 +34,20 @@ public class UserController {
 
 
     @DeleteMapping
-    public ResponseEntity deleteUser(@RequestParam Long id) {
-        try {
-            return ResponseEntity.ok(userService.deleteUser(id));
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body("Ошибочка вышла :(");
+    public ResponseEntity deleteUser(@RequestParam Long id, Authentication authentication) {
+        userPrincipal = (AppUserDetails) authentication.getPrincipal();
+        if (userPrincipal.getId().equals(id) || userPrincipal.getRole().equals("ROLE_ADMIN")) {
+            try {
+                return ResponseEntity.ok(userService.deleteUser(id));
+            } catch (Exception e) {
+                return ResponseEntity.badRequest().body("Ошибочка вышла :(");
+            }
+        } else {
+            return ResponseEntity.badRequest().body("В доступе отказано");
         }
     }
+
+
 }
